+++ src/fmgc/src/guidance/lnav/transitions/PathCaptureTransition.ts
@@ -13,7 +13,7 @@ import { Transition } from '@fmgc/guidance/lnav/Transition';
 import { GuidanceParameters } from '@fmgc/guidance/ControlLaws';
 import { Coordinates } from '@fmgc/flightplanning/data/geo';
 import { Geo } from '@fmgc/utils/Geo';
-import { PathVector, PathVectorType } from '@fmgc/guidance/lnav/PathVector';
+import { PathVector, pathVectorLength, PathVectorType } from '@fmgc/guidance/lnav/PathVector';
 import { CourseChange } from '@fmgc/guidance/lnav/transitions/utilss/CourseChange';
 import { Constants } from '@shared/Constants';
 import { LnavConfig } from '@fmgc/guidance/LnavConfig';
@@ -482,5 +482,28 @@ export class PathCaptureTransition extends Transition {
 
     get repr(): string {
         return `PATH CAPTURE(${this.previousLeg.repr} TO ${this.nextLeg.repr})`;
+    }
+
+    // This is for VNAV to estimate the amount of track miles left
+    getActualDistanceToGo(ppos: LatLongData, trueTrack: number): NauticalMiles {
+        let dtg = 0;
+
+        for (const path of this.predictedPath) {
+            if ('centrePoint' in path) {
+                // Arc
+                const turnSign = this.computedTurnDirection === TurnDirection.Left ? -1 : 1;
+                let trackAngleError = this.computedTargetTrack - trueTrack;
+                if (turnSign !== Math.sign(trackAngleError)) {
+                    trackAngleError += turnSign * 360;
+                }
+
+                dtg += pathVectorLength(path) * trackAngleError / path.sweepAngle;
+            } else if ('endPoint' in path) {
+                // Line
+                dtg += Math.min(pathVectorLength(path), distanceTo(path.endPoint, ppos));
+            }
+        }
+
+        return dtg;
     }
 }