+++ src/fmgc/src/guidance/vnav/wind/WindProfile.ts
@@ -1,0 +1,5 @@
+import { WindComponent } from '@fmgc/guidance/vnav/wind';
+
+export interface WindProfile {
+    getHeadwindComponent(distanceFromStart: NauticalMiles, altitude: Feet, planeHeading: DegreesTrue): WindComponent
+}