+++ src/instruments/src/PFD/instrument.tsx
@@ -157,6 +157,7 @@ class A32NX_PFD extends BaseInstrument {
         this.simVarPublisher.subscribe('engTwoRunning');
         this.simVarPublisher.subscribe('expediteMode');
         this.simVarPublisher.subscribe('setHoldSpeed');
+        this.simVarPublisher.subscribe('tdReached');
         this.simVarPublisher.subscribe('vls');
         this.simVarPublisher.subscribe('trkFpaDeselectedTCAS');
         this.simVarPublisher.subscribe('tcasRaInhibited');
@@ -209,6 +210,13 @@ class A32NX_PFD extends BaseInstrument {
         this.simVarPublisher.subscribe('fac2EstimatedBetaRaw');
         this.simVarPublisher.subscribe('fac1BetaTargetRaw');
         this.simVarPublisher.subscribe('fac2BetaTargetRaw');
+
+        this.simVarPublisher.subscribe('linearDeviationActive');
+        this.simVarPublisher.subscribe('verticalProfileLatched');
+        this.simVarPublisher.subscribe('targetAltitude');
+        this.simVarPublisher.subscribe('showSpeedMargins');
+        this.simVarPublisher.subscribe('upperSpeedMargin');
+        this.simVarPublisher.subscribe('lowerSpeedMargin');
 
         FSComponent.render(<PFDComponent bus={this.bus} instrument={this} />, document.getElementById('PFD_CONTENT'));
     }